+++ src/CMakeLists.txt
@@ -1,10 +1,8 @@
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/ros_messages)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/utils)
+add_subdirectory(ros_messages)
+add_subdirectory(utils)
 
 add_library(rosbag STATIC rosbag.cpp)
-target_include_directories(rosbag PUBLIC ${CMAKE_SOURCE_DIR}/include/utils)
-target_link_libraries(rosbag PUBLIC sensor_msgs parse_util)
+target_link_libraries(rosbag PUBLIC ros_msgs utils)
 
 add_executable(parsePointCloud2Msgs parsePointCloud2Msgs.cpp)
-target_link_libraries(parsePointCloud2Msgs PUBLIC parse_util rosbag)
-target_include_directories(parsePointCloud2Msgs PUBLIC ${CMAKE_SOURCE_DIR}/include/)
+target_link_libraries(parsePointCloud2Msgs PUBLIC rosbag ros_msgs utils)
