+++ src/utils/math.hpp
